refactor: added core module
